pyproject.toml
[project]
name = "store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django>=5.2.2",
    "razorpay>=1.4.2",
    "whitenoise>=6.9.0",
    "django-unfold>=0.59.0",
    "django-environ>=0.12.0",
    "django-thumbnails>=0.8.0",
    "daphne>=4.2.0",
    "django-taggit>=6.1.0",
    "psycopg2>=2.9.10",
    "django-crispy-forms>=2.4",
    "crispy-tailwind>=1.0.3",
    "django-storages[s3]>=1.14.6",
    "boto3>=1.38.31",
    "boto3-stubs>=1.38.31",
    "google-api-python-client>=2.171.0",
]
